Inline ActionMailerHelpers methods and perform cleanup
* remove Gemfile.lock from version control
* use secure connection to rubygems.org
* inline methods from ActionMailerHelpers into the only test that uses them
* alphabetically order some config lines
