+++ lib/generators/clearance/specs/templates/features/clearance/visitor_resets_password_spec.rb
@@ -25,4 +25,20 @@ feature 'Visitor resets password' do
   def page_should_display_change_password_message
     page.should have_content I18n.t('passwords.create.description')
   end
+
+  def mailer_should_have_delivery(recipient, subject, body)
+    ActionMailer::Base.deliveries.should_not be_empty
+
+    message = ActionMailer::Base.deliveries.any? do |email|
+      email.to == [recipient] &&
+        email.subject =~ /#{subject}/i &&
+        email.body =~ /#{body}/
+    end
+
+    message.should be
+  end
+
+  def mailer_should_have_no_deliveries
+    ActionMailer::Base.deliveries.should be_empty
+  end
 end
